requirements.txt
playwright==1.55.0
pytest==8.4.2
allure-pytest==2.15.2
python-dotenv==1.2.1
openpyxl==3.1.5
requests==2.32.5